M9:实体全量接入 patra 审计基座(验收意见⑥)
- 11 实体按语义继承基座家族:聚合根 BaseJpaEntity(campaign/draw/category/
  prompt/generation)、独立子实体 ChildJpaEntity(prize_slot/prompt_like/
  prompt_favorite/ref_image)、值对象表 ValueObjectJpaEntity(generation_
  image/generation_ref);软删变体暂不启用(现有删除语义都是真删)
- 雪花 id 应用层预分配(业务构造器 setId(SnowflakeIdGenerator.getId())),
  数据库自增全部退役;复合主键/@EmbeddedId 三件套删除,成员幂等改走
  UNIQUE(prompt_id,user_id)/UNIQUE(user_id,sha256)
- generation 雪花主键仅内部代理:对外/幂等标识降为 client_task_id 唯一列,
  端口/命令/API 的 id 语义(任务 uuid)不变,涟漪锁死 infra 层;对外零雪花
  Long 暴露,规避 JS 安全整数精度坑
- Flyway V1/V2 就地重写(未部署=绿地):全表审计列组 + BIGINT 主键无默认;
  created_at/updated_at/version 带 DEFAULT 兜底纯 SQL 写入(迁移/收录管线)
- 审计操作人接通:@CurrentUser 解析器把画像挂请求属性,boot 新增
  CurrentUserAuditorConfig(Bean 名必须 auditorAware,starter 按名引用),
  draw/generation 的 created_by 自动=发起用户;无请求上下文留 NULL
- 测试 88→96:审计列自动填充/雪花量级/client_task_id 对外身份/解析器挂
  属性/auditor 三态各有断言;造数 INSERT 补显式 id;改动测试类补 @Timeout
- 规则同步:tech/jpa.md 重写(基座选型表+预分配+auditorAware 命名坑+代理
  键纪律),layers/infra.md、testing/integration-test.md、project-info.md
  随改;旧「不继承 BaseJpaEntity/无雪花」差异条款作废

+++ snb-gallery/snb-gallery-infra/src/main/java/me/supernb/gallery/infra/adapter/persistence/dao/CategoryJpaRepository.java
@@ -6,18 +6,20 @@ import me.supernb.gallery.infra.adapter.persistence.entity.PromptEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-/// gallery.category 仓储。
-public interface CategoryJpaRepository extends JpaRepository<CategoryEntity, Integer> {
+/// category 表 Spring Data 仓储。
+public interface CategoryJpaRepository extends JpaRepository<CategoryEntity, Long> {
 
-    /// 每类目已发布提示词计数(无条目的类目不出现,适配层补 0)。
+    /// 按类目统计已发布条目数(只统计挂了类目的条目)。
     @Query("SELECT p.category.id AS id, COUNT(p) AS cnt FROM PromptEntity p "
             + "WHERE p.status = 'published' AND p.category IS NOT NULL GROUP BY p.category.id")
     List<CategoryCountView> publishedCountByCategory();
 
     /// 类目计数投影。
     interface CategoryCountView {
-        Integer getId();
+        /// 类目 id。
+        Long getId();
 
+        /// 已发布条目数。
         long getCnt();
     }
 }
